Volume option improvements

+++ videoutil.py
@@ -22,7 +22,7 @@ import util
 
 video_formats = ['.mp4', '.wmv', '.mov', '.m4v', '.mpg', '.avi', '.flv']
 
-def ffmpeg_run(pts_in, filters, pts_out, silent = True, expected_length = 0, description = None, bar_pos=None, block=True, line_callback=None):
+def ffmpeg_run(pts_in, filters, pts_out, silent = True, expected_length = 0, description = None, bar_pos=None, block=True, line_callback=None, ignore_errors=False):
     cmd_pts = ['ffmpeg', '-hide_banner -y'] + pts_in
     
     if filters:
@@ -74,7 +74,7 @@ def ffmpeg_run(pts_in, filters, pts_out, silent = True, expected_length = 0, des
                     if 'No decoder surfaces left' in l:
                         p.terminate()
                         error_msg = l
-                    if 'Error' in l:
+                    if not ignore_errors and 'Error' in l:
                         p.terminate()
                         error_msg = l
                     
@@ -120,7 +120,8 @@ def ffprobe_run(pts_in, suppress_errors=True):
     retcode = -1
 
     try:
-        result = subprocess.run(cmd, stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
+        # result = subprocess.run(cmd, stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
+        result = subprocess.run(cmd, stdout=subprocess.PIPE)
         retcode = result.returncode
         if retcode != 0:
             raise Exception("Invalid return code")
@@ -192,338 +193,6 @@ def videos_get(vid_folder, vids_deep, num_vids, num_threads=4):
 
 def get_media_length(input):
     return float(ffprobe_run(['-show_entries format=duration', '-of csv="p=0"', '-i "{}"'.format(input)]))
-
-def videos_analyze_thread(v):
-    start_skip = 10
-    end_skip = 10
-
-    try:
-        result = ffprobe_run([
-            '-select_streams v',
-            '-show_entries format=duration',
-            '-show_entries stream=width,height',
-            '-of csv="p=0"',
-            '-i "{}"'.format(v['video'])
-        ]).split("\n")
-        resolution = list(map(int, result[0].strip().split(',')))
-        length = float(result[1].strip())
-        
-        
-    except BaseException as e:
-        print('Failed analyzing video: {}'.format(v['video']))
-        raise e
-
-    ret = {
-        'full_length': length,
-        'start_at': start_skip,
-        'end_at': length - end_skip,
-        'usable_length': length - start_skip - end_skip,
-        'video': v['video'],
-        'width': resolution[0],
-        'height': resolution[1]
-    }
-        
-    return ret
-    
-def clips_get(videos, beats, fps, volume):
-    frame_time = 1 / fps
-    video_index = 0
-    beat_clips = []
-    framenr = 0
-
-    output_length = beats[-1]
-
-    for v in videos:
-        v['clips'] = []
-    
-    for beat_index in util.Utqdm(range(len(beats) - 1), desc="Clip assignment"):
-        beat_start = beats[beat_index]
-        beat_end = beats[beat_index + 1]
-        beat_duration = beat_end - beat_start
-
-        last_frame_time = framenr * frame_time
-        missing_frames = math.floor((beat_end - last_frame_time) / frame_time)
-
-        if video_index >= len(videos):
-            video_index = 0
-            random.shuffle(videos)
-
-        video_state = videos[video_index]
-        video_index += 1
-
-        beat_pos = beat_start / output_length
-
-        found = False
-        tries = len(videos)
-        while not found:
-            if tries <= 0:
-                print("Failed finding clip for beat: {} - {}".format(beat_start, beat_end))
-                return False
-
-            beat_pos_osffset = util.clamp(beat_pos + (random.random() * 0.2 - 0.1), 0, 1)
-
-            clip_start = video_state['start_at'] + (video_state['usable_length'] * beat_pos_osffset)
-            clip_end = clip_start + beat_duration
-
-            if clip_end > video_state['full_length']:
-                video_index += 1
-                if video_index >= len(videos):
-                    video_index = 0
-                video_state = videos[video_index]
-                tries -= 1
-                continue
-            
-            for c in video_state['clips']:
-                if c['clip_start'] >= clip_start and c['clip_end'] <= clip_start:
-                    video_index += 1
-                    if video_index >= len(videos):
-                        video_index = 0
-                    video_state = videos[video_index]
-                    tries -= 1
-                    continue
-
-                if c['clip_start'] >= clip_end and c['clip_end'] <= clip_end:
-                    video_index += 1
-                    if video_index >= len(videos):
-                        video_index = 0
-                    video_state = videos[video_index]
-                    tries -= 1
-                    continue
-            
-            found = True
-                
-
-        clip = {
-            'index': beat_index,
-
-            'beat_start': beat_start,
-            'beat_end': beat_end,
-            'duration': beat_duration,
-            
-            'video': video_state['video'],
-            'video_state': video_state,
-            'clip_start': clip_start,
-            'clip_end': clip_end,
-            
-            'frame': framenr,
-            'framecount': missing_frames,
-            'volume': volume
-        }
-
-        beat_clips.append(clip)
-        video_state['clips'].append(clip)
-        framenr += missing_frames
-        
-        util.handle_tqdm_out()
-        
-        # TODO scene detection
-    
-    return beat_clips
-
-def batch(iterable, n=1):
-    l = len(iterable)
-    for ndx in range(0, l, n):
-        yield iterable[ndx:min(ndx + n, l)]
-
-def clips_generate_batched(beat_clips, fps, resolution, bitrate='3M', batch_size=10, num_threads=4, cuda=False, pre_seek=0):
-    videos_file_name = util.get_tmp_file('txt')
-    videos_file = open(videos_file_name, 'w')
-    
-    frame_time = 1 / fps
-    video_clips = {}
-    for i,b in enumerate(beat_clips):
-        videos_file.write("file '{}.mp4'\n".format(i))
-
-        if b['video'] not in video_clips:
-            video_clips[b['video']] = []
-        video_clips[b['video']].append(b)
-
-    batches = []
-
-    for v,all_clips in video_clips.items():
-        all_clips.sort(key=lambda x: x['clip_start'])
-        for clips in batch(all_clips, batch_size):
-            
-            #can_cuda = cuda
-            #if can_cuda and clips[0]['video_state']['width'] > 4096:
-            #    can_cuda = False
-
-            batches.append({
-                'video': v,
-                'clips': clips,
-                'fps': fps,
-                'resolution': resolution,
-                'bitrate': bitrate,
-                'cuda': cuda,
-                'pre_seek': pre_seek
-            })
-
-    videos_file.close()
-
-    threaded = True
-        
-    with util.Utqdm(total=len(batches), desc="Splitting videos") as pbar:
-        futures = []
-
-        def callback(result):
-            pbar.update()
-
-
-        if not threaded:
-            for c in batches:
-                callback(clips_generate_batched_thread(c))
-                pbar.update()
-                return False
-        else:
-
-            with ThreadPoolExecutor(max_workers=num_threads) as executor:
-                
-                shared_vars = {}
-
-                if cuda:
-                    shared_vars['cuda_sessions'] = 0
-
-                for c in batches:
-                    future = executor.submit(clips_generate_batched_thread, c, shared_vars)
-                    future.add_done_callback(callback)
-                    futures.append(future)
-
-                for future in futures:
-                    try:
-                        res = future.result()
-                    except BaseException as e:
-                        for f in futures:
-                            f.cancel()
-                        raise e
-
-    return videos_file_name
-
-
-def clips_generate_batched_thread(myargs, shared_vars):
-    cmd_in = []
-    cmd_out = []
-
-    cuda_dec = myargs['cuda']
-    cuda_enc = myargs['cuda']
-
-    if cuda_enc and shared_vars['cuda_sessions'] >= 3:
-        cuda_enc = False
-    if cuda_dec and myargs['clips'][0]['video_state']['width'] > 4096:
-        cuda_dec = False
-
-    if not cuda_enc and not cuda_dec:
-       myargs['cuda'] = False 
-    
-    if cuda_enc:
-        shared_vars['cuda_sessions'] = shared_vars['cuda_sessions'] + 1
-
-    if cuda_dec:
-        cmd_in += ['-hwaccel cuda', '-hwaccel_output_format cuda']
-    if myargs['cuda']:
-        cmd_in.append('-extra_hw_frames 12')
-
-    for i,c in enumerate(myargs['clips']):
-        cmd_in += [
-            '-ss {}'.format(timestamp(max(0, c['clip_start'] - myargs['pre_seek']))),
-            '-t {}'.format(c['duration'] + myargs['pre_seek'] + 0.5),
-            '-i "{}"'.format(myargs['video']),
-        ]
-
-        if myargs['pre_seek']:
-            cmd_in.append('-ss {}'.format(timestamp(myargs['pre_seek'])))
-
-        filters = [
-            'fps={}'.format(myargs['fps'])
-        ]
-
-        if myargs['cuda']:
-            if not cuda_dec:
-                filters.append('format=nv12,hwupload_cuda')
-
-            filters.append('scale_cuda={}:force_original_aspect_ratio=1'.format(myargs['resolution']))
-
-            if not cuda_enc:
-                filters.append('hwdownload,format=nv12')
-                
-            #'hwdownload,format=nv12',
-            #'pad={}:(ow-iw)/2:(oh-ih)/2'.format(myargs['resolution'])
-            
-        else:
-            filters += [
-                'scale={}:force_original_aspect_ratio=1'.format(myargs['resolution']),
-                'pad={}:(ow-iw)/2:(oh-ih)/2'.format(myargs['resolution'])
-            ]
-
-        cmd_out += [
-            '-vf "{}"'.format(','.join(filters)),
-            '-map {}:v'.format(i),
-            '-vframes {}'.format(c['framecount']),
-            #'-vsync cfr',
-            #'-video_track_timescale 60000'
-        ]
-
-        if c['volume'] > 0:
-            cmd_out += [
-                '-af atrim=duration={}'.format((1/myargs['fps'])*c['framecount']),
-                '-map {}:a'.format(i),
-                '-c:a aac',
-                '-b:a 192k'
-            ]
-
-        if cuda_enc:
-            cmd_out.append('-c:v h264_nvenc')
-            
-        cmd_out +=[
-            '-b:v {}'.format(myargs['bitrate']),
-            '{}/{}.mp4'.format(util.get_tmp_dir(), c['index'])
-        ]
-    
-    ffmpeg_run(cmd_in, None, cmd_out, True)
-
-    for c in myargs['clips']:
-        clip_out = '{}/{}.mp4'.format(util.get_tmp_dir(), c['index'])
-        if not os.path.isfile(clip_out):
-            raise Exception("Failed splitting {}, {} was not found".format(myargs['video'], clip_out))
-
-        framecount = int(ffprobe_run([
-            '-select_streams v:0',
-            '-count_packets',
-            '-show_entries stream=nb_read_packets',
-            '-of csv=p=0',
-            '-i "{}"'.format(clip_out)]
-        ))
-
-        if framecount != c['framecount']:
-            raise Exception("{} has {} frames instead of the requestes {}".format(clip_out, framecount, c['framecount']))
-
-    if cuda_enc:
-        shared_vars['cuda_sessions'] = max(0, shared_vars['cuda_sessions'] - 1)
-    
-    return myargs
-
-
-def clips_merge(output, audio, vids_file, expected_length):
-    cmd_in = [
-        '-f concat',
-        '-i "{}"'.format(vids_file)
-    ]
-
-    if audio:
-        cmd_in.append('-i "{}"'.format(audio))
-
-    filters = []
-    cmd_out = ['-map 0:v']
-    
-    if audio:
-        cmd_out.append('-map 1:a')
-        cmd_out.append('-c:a aac')
-
-    # cmd_out.append('-c copy')
-    # cmd_out.append('-c:v h264_nvenc')
-    
-    cmd_out.append('"{}"'.format(output))
-
-    ffmpeg_run(cmd_in, filters, cmd_out, expected_length=expected_length, description="Merging clips together", block=True)
 
 def timestamp(seconds):
     minutes = seconds // 60
@@ -634,7 +303,7 @@ def apply_circles(beats, video, keep_audio, output, expected_length = 0, bar_pos
         vnum += 1
         
     pts_out = [
-        # '-c:v h264_nvenc',
+        '-c:v {}'.format(util.video_ctx.video_codec),
         '-map [v{}]'.format(vnum)
     ]
